chore(nlp): Add a new level for some packages.
